ai-web-summarizer/requirements.txt
requests
beautifulsoup4
python-dotenv
google-generativeai
IPython
